+++ SocialNews/src/main/java/it/unipi/lsmsd/socialnews/dao/implement/ReportDAOImpl.java
@@ -19,11 +19,11 @@ public class ReportDAOImpl implements ReportDAO {
      * Inserts new report into the database
      *
      * @param report report object containing information of the new report
-     * @return number of created report
+     * @return identifier of created report
      * @throws SocialNewsDataAccessException in case of failure of the insert operation on database
      */
     @Override
-    public Integer addReport(Report report) throws SocialNewsDataAccessException {
+    public Long addReport(Report report) throws SocialNewsDataAccessException {
         return neo4JReportDAO.addReport(report);
     }
